VERSION PRODUCCION

+++ boleta-ejb/src/main/java/sv/gob/mined/boleta/ejb/EnvioDeBoletasFacade.java
@@ -61,7 +61,7 @@ public class EnvioDeBoletasFacade {
         int docenteNoEncontrados = 0;
         int correosNoEnviados = 0;
 
-        Long idCodigoGenerado = persistenciaFacade.getPkCodigoGeneradoByCodDepaAndMesAnho(codDepa, mesAnho);
+        Integer idCodigoGenerado = persistenciaFacade.getPkCodigoGeneradoByCodDepaAndMesAnho(codDepa, mesAnho);
 
         SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy - HH:mm:ss");
 
@@ -176,7 +176,7 @@ public class EnvioDeBoletasFacade {
             folderNoEncontrado.mkdir();
         }
 
-        Long idCodigoGenerado = persistenciaFacade.getPkCodigoGeneradoByCodDepaAndMesAnho(codDepa, mesAnho);
+        Integer idCodigoGenerado = persistenciaFacade.getPkCodigoGeneradoByCodDepaAndMesAnho(codDepa, mesAnho);
 
         SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy - HH:mm:ss");
